Reorganize MyListView
