+++ src/lib/views/MyListView/index.ts
@@ -1,0 +1,3 @@
+import MyListView from './MyListView.svelte';
+
+export default MyListView;
